ENH: Further WIP for the runner
Now saves tasklists as runner prefs
Has shortcut keys to switch to runner
Downloads and runs local PsychoJS lib for debugging
Runner has menu for a few tasks
Refactor changes what is stored in expCtrl

Needs documenting, testing, and todos

+++ psychopy/app/runner/runner.py
@@ -3,6 +3,7 @@ import os
 import sys
 import time
 from pathlib import Path
+import requests
 import webbrowser
 import xml.etree.ElementTree as xml
 from subprocess import Popen, PIPE
@@ -10,13 +11,13 @@ from subprocess import Popen, PIPE
 from psychopy.app import icons
 from psychopy.constants import PY3
 from psychopy.app.stdOutRich import StdOutRich
-from psychopy.app.builder.builder import BuilderFrame
 from psychopy.app.coder.coder import CoderFrame
+from psychopy.app.builder.builder import BuilderFrame
 from psychopy.projects.pavlovia import getProject
 
 
 # todo: Make print statements ALERTS
-# todo: Set menu to allow app quit, view of Builder, Coder, and allow opening of listed tasks in those frames.
+# todo: store server processes, to end them after a timeout, or end them automatically on each run
 
 class RunnerFrame(wx.Frame):
     """Defines construction of the Psychopy Runner Frame"""
@@ -29,23 +30,75 @@ class RunnerFrame(wx.Frame):
                                           style=wx.DEFAULT_FRAME_STYLE,
                                           name=title,
                                           )
+
+        # Create menu
+        self.runnerMenu = wx.MenuBar()
+        self.makeMenu()
+        self.SetMenuBar(self.runnerMenu)
+
         self.app = app
         self.frameType = 'runner'
         self.app.trackFrame(self)
+
         self.panel = RunnerPanel(self, id, title, app)
 
         self.mainSizer = wx.BoxSizer(wx.VERTICAL)
         self.mainSizer.Add(self.panel, 1, wx.EXPAND | wx.ALL)
         self.SetSizerAndFit(self.mainSizer)
+        self.prefs = self.app.prefs.runner
+        self.loadTasks()
 
         self.Bind(wx.EVT_CLOSE, self.onClose)
 
-    def addExperiment(self, fileName=None):
+    def addTask(self, evt=None, fileName=None):
         self.panel.addExperiment(fileName=fileName)
+
+    def removeTask(self, evt=None):
+        self.panel.removeTask(evt)
 
     @property
     def stdOut(self):
         return self.panel.stdoutCtrl
+
+    def makeMenu(self):
+        """
+        Create Runner menu.
+        """
+        # File Menu
+        fileMenu = wx.Menu()
+
+        items = [{'id': wx.ID_ADD, 'label': 'Add task', 'status': 'Adding task...', 'func': self.addTask},
+                 {'id': wx.ID_REMOVE, 'label': 'Remove task', 'status': 'Removing task...', 'func': self.removeTask},
+                 {'id': wx.ID_CLEAR, 'label': 'Clear all', 'status': 'Clearing tasks...', 'func': self.clearTasks},
+                 {'id': wx.ID_SAVE, 'label': 'Save task', 'status': 'Saving task...', 'func': self.saveTasks},
+                 {'id': wx.ID_COPY, 'label': 'Load task', 'status': 'Loading task...', 'func': self.loadTasks},
+                 {'id': wx.ID_CLOSE_FRAME, 'label': 'Close', 'status': 'Closing Runner...', 'func': self.onClose},
+                 {'id': wx.ID_EXIT, 'label': 'Quit', 'status': 'Quitting PsychoPy...', 'func': self.onQuit},
+                 ]
+
+        separators = ['clear all', 'load task']
+        for item in items:
+            fileItem = fileMenu.Append(item['id'], item['label'], item['status'])
+            self.Bind(wx.EVT_MENU, item['func'], fileItem)
+            if item['label'].lower() in separators:
+                fileMenu.AppendSeparator()
+
+        # View Menu
+        viewMenu = wx.Menu()
+        viewItems = ['Builder', 'Coder']
+        viewFunctions = [self.viewBuilder, self.viewCoder]
+
+        for idx, item in enumerate(viewItems):
+            menuItem = wx.MenuItem(parentMenu=viewMenu,
+                                   id=idx,
+                                   text='View {}'.format(item),
+                                   helpString='Open a {} frame'.format(item),
+                                   )
+            viewMenu.Append(menuItem)
+            self.Bind(wx.EVT_MENU, viewFunctions[idx], id=idx)
+
+        self.runnerMenu.Append(fileMenu, 'File')
+        self.runnerMenu.Append(viewMenu, 'View')
 
     def onURL(self, evt):
         """
@@ -66,6 +119,17 @@ class RunnerFrame(wx.Frame):
             print("Could not open URL: {}".format(evt.String))
         wx.EndBusyCursor()
 
+    def saveTasks(self, evt=None):
+        self.prefs['taskList'] = self.taskList
+        self.app.prefs.saveUserPrefs()
+
+    def loadTasks(self, evt=None):
+        for filePath in self.prefs['taskList']:
+            self.addTask(fileName=filePath)
+
+    def clearTasks(self, evt=None):
+        self.panel.expCtrl.DeleteAllItems()
+
     def onClose(self, evt):
         """
         Defines Frame closing behavior.
@@ -73,9 +137,28 @@ class RunnerFrame(wx.Frame):
         """
         lastFrame = len(self.app.getAllFrames()) == 1
         if lastFrame:
-            self.Destroy()  # required
-            self.app.forgetFrame(self)
-            self.app.quit(evt)
+            self.onQuit()
+        else:
+            self.Hide()
+
+    def onQuit(self, evt=None):
+        self.saveTasks()
+        self.Destroy()  # required
+        self.app.forgetFrame(self)
+        self.app.quit(evt)
+
+    def viewBuilder(self, evt):
+        self.app.showBuilder(fileList=[str(self.panel.currentFile)])
+
+    def viewCoder(self, evt):
+        self.app.showCoder(fileList=[str(self.panel.currentFile)])
+
+    @property
+    def taskList(self):
+        temp = []
+        for idx in range(self.panel.expCtrl.GetItemCount()):
+            temp.append(self.panel.expCtrl.GetItem(idx, 1).Text)
+        return temp
 
 
 class RunnerPanel(wx.Panel):
@@ -93,9 +176,11 @@ class RunnerPanel(wx.Panel):
 
         self.app = app
         self.parent = parent
-        self.taskList = {}
         self.localProcess = None
         self.serverProcess = None
+
+        self.currentFile = None
+        self.currentProject = None
         self.currentSelection = None
 
         # Set ListCtrl for list of tasks
@@ -111,16 +196,13 @@ class RunnerPanel(wx.Panel):
 
         self.expCtrl.Bind(wx.EVT_LIST_ITEM_SELECTED, self.onItemSelected, self.expCtrl)
         self.expCtrl.InsertColumn(0, 'File')
-        self.expCtrl.InsertColumn(1, 'Type')
-        self.expCtrl.InsertColumn(2, 'Path')
-        self.expCtrl.InsertColumn(3, 'Project')
-        self.expCtrl.SetColumnWidth(1, 75)
+        self.expCtrl.InsertColumn(1, 'Path')
 
         # Set buttons
         plusBtn = self.makeBmpButton(main='addExp32.png')
         negBtn = self.makeBmpButton(main='removeExp32.png')
         runLocalBtn = self.makeBmpButton(main='run32.png')
-        stopTaskBtn = self.makeBmpButton(main='stop32.png')
+        self.stopBtn = stopTaskBtn = self.makeBmpButton(main='stop32.png')
         onlineBtn = self.makeBmpButton(main='globe32.png', emblem='run16.png')
         onlineDebugBtn = self.makeBmpButton(main='globe32.png', emblem='bug16.png')
 
@@ -133,30 +215,31 @@ class RunnerPanel(wx.Panel):
 
         # Bind events to buttons
         self.Bind(wx.EVT_BUTTON, self.addExperiment, plusBtn)
-        self.Bind(wx.EVT_BUTTON, self.removeExperiment, negBtn)
+        self.Bind(wx.EVT_BUTTON, self.removeTask, negBtn)
         self.Bind(wx.EVT_BUTTON, self.runLocal, runLocalBtn)
         self.Bind(wx.EVT_BUTTON, self.stopTask, stopTaskBtn)
         self.Bind(wx.EVT_BUTTON, self.runOnline, onlineBtn)
         self.Bind(wx.EVT_BUTTON, self.runOnlineDebug, onlineDebugBtn)
 
         # GridBagSizer
-        gridSize = 4
+        gridRow, gridCol = (4,2)
         self.expBtnSizer = wx.GridBagSizer(vgap=0, hgap=0)
 
         # Add ListCtrl
-        self.expBtnSizer.Add(self.expCtrl, (0, 0), (gridSize, gridSize), wx.EXPAND)
+        self.expBtnSizer.Add(self.expCtrl, (0, 0), (gridRow, gridCol), wx.EXPAND)
 
         # Add buttons
-        self.expBtnSizer.Add(plusBtn, (0,4), (1,1), wx.ALL, 5)
-        self.expBtnSizer.Add(negBtn, (1,4), (1,1), wx.ALL, 5)
-        self.expBtnSizer.Add(runLocalBtn, (3,4),(1,1), wx.ALL | wx.EXPAND, 5)
-        self.expBtnSizer.Add(stopTaskBtn, (3, 5), (1, 1), wx.ALL | wx.EXPAND, 5)
-        self.expBtnSizer.Add(onlineBtn, (3,6),(1,1), wx.ALL | wx.EXPAND, 5)
-        self.expBtnSizer.Add(onlineDebugBtn, (3,7),(1,1), wx.ALL | wx.EXPAND, 5)
+        self.expBtnSizer.Add(plusBtn, (0,2), (1,1), wx.ALL, 5)
+        self.expBtnSizer.Add(negBtn, (1,2), (1,1), wx.ALL, 5)
+        self.expBtnSizer.Add(runLocalBtn, (3,2),(1,1), wx.ALL | wx.EXPAND, 5)
+        self.expBtnSizer.Add(stopTaskBtn, (3, 3), (1, 1), wx.ALL | wx.EXPAND, 5)
+        self.expBtnSizer.Add(onlineBtn, (3,4),(1,1), wx.ALL | wx.EXPAND, 5)
+        self.expBtnSizer.Add(onlineDebugBtn, (3,5),(1,1), wx.ALL | wx.EXPAND, 5)
 
-        for idx in range(gridSize):
+        for idx in range(gridRow):
             self.expBtnSizer.AddGrowableRow(idx)
-            self.expBtnSizer.AddGrowableCol(idx)
+            if idx < 2:
+                self.expBtnSizer.AddGrowableCol(idx)
 
         # Set main sizer
         self.mainSizer = wx.BoxSizer(wx.VERTICAL)
@@ -197,39 +280,63 @@ class RunnerPanel(wx.Panel):
         """
         Kills script processes currently running.
         """
-        # Local script
-        if self.localProcess is not None:
-            self.localProcess.stopFile(evt)
-            self.localProcess = None
-            print("***** Experiment quit. *****")
+        # Check whether script ended automatically
+        if self.processExists and not (evt.EventObject.ClassName == 'wxBitmapButton'):
+            return
 
-        # Subprocess script running local server
+        # Stop Runner script local process
+        if self.localProcess is not None:
+            self.Bind(wx.EVT_IDLE, None)
+            try:
+                self.localProcess.stopFile(evt)
+            except TypeError:
+                pass  # coder Process already dead
+            self.localProcess = None
+            print("***** Experiment finished. *****")
+
+        # Stop subprocess script running local server
         if self.serverProcess is not None:
             self.serverProcess.kill()
             self.serverProcess = None
             print("***** Local server shut down. *****")
 
+        # Stop Builder or Coder scripts
+        for thisFrame in self.app.getAllFrames():
+            if hasattr(thisFrame, 'scriptProcess') and thisFrame.scriptProcess is not None:
+                thisFrame.stopFile()
+                print("***** Experiment finished. *****")
+
     def runLocal(self, evt):
         """
-        Run experiment from Builder or Coder Frames, depending on file type.
+        Run experiment from Builder - Builder can run both .py and .psyexp filetypes,
+        if builderFrame.filename is set after frame creation.
         """
-        if self.currentSelection is None:
+        if self.currentSelection is None or self.localProcess is not None:
             return
 
-        fileType = self.expCtrl.GetItem(self.currentSelection, 1).Text
-        fileName = self.expCtrl.GetItem(self.currentSelection, 2).Text
-
-        if fileType == '.py':
-            title = "PsychoPy3 Coder (IDE) (v{})".format(self.app.version)
-            self.localProcess = CoderFrame(None, -1,
-                                   title=title,
-                                   files=[fileName], app=self.app)
-        else:
+        if self.currentFile.suffix == '.psyexp':
             title = "PsychoPy3 Experiment Builder (v{})".format(self.app.version)
             self.localProcess = BuilderFrame(None, -1,
-                                     title=title,
-                                     fileName=fileName, app=self.app)
+                                             title=title,
+                                             fileName=str(self.currentFile),
+                                             app=self.app)
+        else:
+            title = "PsychoPy3 Experiment Builder (v{})".format(self.app.version)
+            self.localProcess = CoderFrame(None, -1,
+                                           title=title,
+                                           files=[str(self.currentFile)],
+                                           app=self.app)
         self.localProcess.runFile()
+        time.sleep(.5)
+        self.Bind(wx.EVT_IDLE, self.stopTask)
+
+    @property
+    def processExists(self):
+        if not hasattr(self.localProcess, 'scriptProcess'):
+            return
+        if self.localProcess.scriptProcess is not None:
+            time.sleep(.1)
+            return True
 
     def runOnline(self, evt):
         """
@@ -238,15 +345,12 @@ class RunnerPanel(wx.Panel):
         if self.currentSelection is None:
             return
 
-        currentFile = self.expCtrl.GetItem(self.currentSelection, 2).Text
-        currentProj = self.expCtrl.GetItem(self.currentSelection, 3).Text
-
-        if currentProj not in [None, "None", ''] and Path(currentFile).suffix == '.psyexp':
+        if self.currentProject not in [None, "None", ''] and self.currentFile.suffix == '.psyexp':
             webbrowser.open(
                 "https://pavlovia.org/run/{}/{}"
-                    .format(currentProj,
+                    .format(self.currentProject,
                             self.outputPath(
-                                currentFile)))
+                                self.currentFile)))
 
     def runOnlineDebug(self, evt, port=12002):
         """
@@ -261,16 +365,15 @@ class RunnerPanel(wx.Panel):
         if self.currentSelection is None:
             return
 
-        currentFile = self.expCtrl.GetItem(self.currentSelection, 2).Text
-        currentProj = self.expCtrl.GetItem(self.currentSelection, 3).Text
-        htmlPath = Path(currentFile).parent / self.outputPath(Path(currentFile))
+        self.getPsychoJS()
+        htmlPath = self.currentFile.parent / self.outputPath(self.currentFile)
         server = ["SimpleHTTPServer", "http.server"][PY3]
         pythonExec = Path(sys.executable)
-        cmd = [pythonExec, "-m", server, str(port)]
+        command = [str(pythonExec), "-m", server, str(port)]
 
-        if currentProj not in [None, "None", ''] and Path(currentFile).suffix == '.psyexp':
+        if self.currentProject not in [None, "None", ''] and self.currentFile.suffix == '.psyexp':
             if self.serverProcess is None:
-                self.serverProcess = Popen(cmd,
+                self.serverProcess = Popen(command,
                                            bufsize=1,
                                            cwd=htmlPath,
                                            stdout=PIPE,
@@ -286,6 +389,19 @@ class RunnerPanel(wx.Panel):
 
     def onURL(self, evt):
         self.parent.onURL(evt)
+
+    def getPsychoJS(self):
+        libPath = self.currentFile.parent / 'lib'
+        if not os.path.exists(libPath):
+            os.makedirs(libPath)
+
+        ver = '.'.join(self.app.version.split('.')[:2])
+        psychoJSLibs = ['core', 'data', 'util', 'visual', 'sound']
+        for lib in psychoJSLibs:
+            url = "https://lib.pavlovia.org/{}-{}.js".format(lib, ver)
+            req = requests.get(url)
+            with open(libPath / "{}-{}.js".format(lib, ver), 'wb') as f:
+                f.write(req.content)
 
     def addExperiment(self, evt=None, fileName=None):
         """
@@ -320,39 +436,43 @@ class RunnerPanel(wx.Panel):
             if self.expCtrl.FindItem(-1, temp.name) > -1:
                 continue
 
-            # Check for project
-            project = getProject(str(temp))
-            if hasattr(project, 'id'):
-                project = project.id
-
             # Set new item in listCtrl
             index = self.expCtrl.InsertItem(self.expCtrl.GetItemCount(), str(temp.name))
-            self.expCtrl.SetItem(index, 1, str(temp.suffix))
-            self.expCtrl.SetItem(index, 2, str(temp))
-            self.expCtrl.SetItem(index, 3, str(project))
+            self.expCtrl.SetItem(index, 1, str(temp))
+
+            # Set item selection
+            self.expCtrl.SetItemState(self.currentSelection or 0, 0, wx.LIST_STATE_SELECTED)
+            self.expCtrl.Select(self.expCtrl.GetItemCount() - 1)
 
         # Set column width
         self.expCtrl.SetColumnWidth(0, wx.LIST_AUTOSIZE)
-        self.expCtrl.SetColumnWidth(2, wx.LIST_AUTOSIZE)
-        self.expCtrl.SetColumnWidth(3, wx.LIST_AUTOSIZE)
+        self.expCtrl.SetColumnWidth(1, wx.LIST_AUTOSIZE)
 
-        # Set item selection
-        self.expCtrl.SetItemState(self.currentSelection or 0, 0, wx.LIST_STATE_SELECTED)
-        self.expCtrl.Select(self.expCtrl.GetItemCount() - 1)
-
-    def removeExperiment(self, evt):
+    def removeTask(self, evt):
         """
         Removes experiment entry from the expList listctrl.
         """
+        if self.currentSelection is None:
+            return
+
         self.expCtrl.DeleteItem(self.currentSelection)
         if self.expCtrl.GetItemCount() == 0:
             self.currentSelection = None
+            self.currentProject = None
+            self.currentFile = None
 
     def onItemSelected(self, evt):
         """
         Sets currentSelection to index of currently selected list item.
         """
         self.currentSelection = evt.Index
+        self.currentFile = Path(self.expCtrl.GetItem(self.currentSelection, 1).Text)
+        self.currentProject = None
+
+        # Check for project
+        project = getProject(str(self.currentFile))
+        if hasattr(project, 'id'):
+            self.currentProject = project.id
 
     def outputPath(self, filePath):
         """
